MBX-3390 Some refactor
