+++ MindboxTests/InApp/Tests/InAppConfigResponseTests/InappMapperTests.swift
@@ -1,0 +1,17 @@
+//
+//  InappManagerTests.swift
+//  MindboxTests
+//
+//  Created by vailence on 14.11.2024.
+//  Copyright © 2024 Mindbox. All rights reserved.
+//
+
+import Testing
+
+struct InappMapperTests {
+
+    @Test func <#test function name#>() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
